app/external/ml_orchestrator/requirements.txt
transformers>=4.44.0
accelerate>=0.33.0
torch>=2.2.0
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0
hdbscan>=0.8.36
joblib>=1.4.2
scikit-learn>=1.4.2
tqdm>=4.66.4
python-dotenv>=1.0.1
ujson>=5.10.0
streamlit>=1.37.0
openai>=1.40.0
